Mobile: Start edit from top
This hopefully alleviates some of the issue in
https://github.com/laurent22/joplin/issues/1838. At least for me I don't really
need a way to scroll quickly, but rather I usually want to add to the top.

This commit makes the mobile experience more consistent both with the desktop
experience and internally. In the desktop version notes are shown from the top,
not the bottom and on mobile the "view note" view already defaults to showing
the top of the note and in my opinion it's somewhat surprising when changing to
edit that it suddenly jumps to the end.